Integrate additional pattern regressions into existing tests
